feat!: Refactor LocalSourceSaveOptions
